backend/requirements.txt
fastapi==0.108.0
uvicorn==0.25.0
python-multipart==0.0.6
python-dotenv==1.0.0
supabase==2.0.3
pydantic==2.5.2 